requirements.txt
Appium-Python-Client>=4.0.0
mitmproxy>=10.0.0
requests>=2.31.0
httpx>=0.27.0
frida-tools>=12.0.0
objection>=1.11.0
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0
pydantic>=2.0.0
rich>=13.0.0
